// backend/src/types/MachineryMaintenance/refuelingRecord.dto.ts
export type CreateRefuelingRecordDTO = {

    maquinariaId: number;
    fecha: string;
    litros: number;
    precio: number;
    operador: string;

}

export type UpdateRefuelingRecordDTO ={

    maquinariaId: number;
    fecha: string;
    litros: number;
    precio: number;
    operador: string;

}